move ClassifyHeaderText into main.py as ClassifyMainHeader
