+++ WebSearcher/classifiers/__init__.py
@@ -1,9 +1,7 @@
 from .footer import ClassifyFooter
-from .header_text import ClassifyHeaderText
 from .main import ClassifyMain
 
 __all__ = [
     "ClassifyFooter",
-    "ClassifyHeaderText",
     "ClassifyMain",
 ]
